refactor(automation): extract cron utils
